requirements.txt
discord.py
pytz
flask
pillow
